Add tests for status 400
