+++ server/controllers/controllers.js
@@ -64,6 +64,11 @@ const getAll = async (req, res) => {
 
   if (data) {
     res.status(200).send(generateGeoData(data.rows));
+  } else {
+    res.status(400).send({
+      data: 'No data',
+      success: false,
+    });
   }
 };
 
@@ -75,6 +80,11 @@ const getZipLayer = async (req, res) => {
 
   if (data) {
     res.status(200).send(generateZipData(data.rows));
+  } else {
+    res.status(400).send({
+      data: 'No data',
+      success: false,
+    });
   }
 };
 
@@ -90,6 +100,11 @@ const getPointData = async (req, res) => {
 
   if (data) {
     res.status(200).send(data.rows);
+  } else {
+    res.status(400).send({
+      data: 'No data',
+      success: false,
+    });
   }
 };
 
@@ -117,6 +132,11 @@ const getTimestamps = async (req, res) => {
 
   if (data) {
     res.status(200).send(generateGeoData(data.rows));
+  } else {
+    res.status(400).send({
+      data: 'No data',
+      success: false,
+    });
   }
 };
 
@@ -143,6 +163,11 @@ const drawSelect = async (req, res) => {
 
   if (data) {
     res.status(200).send(generateGeoData(data.rows));
+  } else {
+    res.status(400).send({
+      data: 'No data',
+      success: false,
+    });
   }
 };
 
@@ -164,6 +189,11 @@ const zipSelect = async (req, res) => {
 
   if (data) {
     res.status(200).send(generateGeoData(data.rows));
+  } else {
+    res.status(400).send({
+      data: 'No data',
+      success: false,
+    });
   }
 };
 
@@ -193,6 +223,11 @@ const graph = async (req, res) => {
 
   if (data) {
     res.status(200).send(data.rows);
+  } else {
+    res.status(400).send({
+      data: 'No data',
+      success: false,
+    });
   }
 };
 
@@ -217,6 +252,11 @@ const zipGraph = async (req, res) => {
 
   if (data) {
     res.status(200).send(data.rows);
+  } else {
+    res.status(400).send({
+      data: 'No data',
+      success: false,
+    });
   }
 };
